chore (refs T22614): move DpCheckboxGroup to client/js.
